pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "screen-time-saver"
version = "0.1.0"
description = "AI-powered social media summarizer — turn 90 minutes of scrolling into a 10-minute podcast"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.39.0",
    "edge-tts>=6.1.0",
    "yt-dlp>=2024.0.0",
    "feedparser>=6.0.0",
    "pydantic>=2.0.0",
    "typer[all]>=0.9.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21"]

[project.scripts]
screen-time-saver = "screen_time_saver.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
